rename logic.rs to logic_manager.rs
