+++ src/logic/logic_manager.rs
